pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dock-cli"
version = "0.2.0"
description = "Manage macOS Dock from YAML configuration"
authors = [{name = "James Sawle"}]
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]
readme = "README.md"

[project.scripts]
dock = "dock.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["dock"]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.10",
    "mypy>=1.0",
    "ruff>=0.1.0",
    "types-pyyaml>=6.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
